+++ R/crew_worker.R
@@ -1,9 +1,9 @@
 #' @title Crew worker.
 #' @export
 #' @family utilities
-#' @description Launch a `crew` worker which runs a `mirai` server.
+#' @description Launch a `crew` worker which runs a `mirai` daemon.
 #' @return `NULL` (invisibly)
-#' @param settings Named list of arguments to `mirai::server()`.
+#' @param settings Named list of arguments to `mirai::daemon()`.
 #' @param launcher Character of length 1, name of the launcher.
 #' @param worker Positive integer of length 1, index of the worker.
 #'   This worker index remains the same even when the current instance
@@ -19,5 +19,5 @@ crew_worker <- function(settings, launcher, worker, instance) {
     CREW_WORKER = worker,
     CREW_INSTANCE = instance
   )
-  do.call(what = mirai::server, args = settings)
+  do.call(what = mirai::daemon, args = settings)
 }