+++ tests/testthat/test-crew_launcher_local.R
@@ -34,7 +34,7 @@ crew_test("crew_launcher_local() can run a task on a worker", {
   crew::crew_retry(
     ~{
       daemons <- rlang::duplicate(
-        x = mirai::daemons(.compute = client$name)$daemons,
+        x = mirai::status(.compute = client$name)$daemons,
         shallow = FALSE
       )
       if (is.null(nrow(daemons))) {
@@ -120,7 +120,7 @@ crew_test("crew_launcher_local() can run a task and time out a worker", {
   crew::crew_retry(
     ~{
       daemons <- rlang::duplicate(
-        x = mirai::daemons(.compute = client$name)$daemons,
+        x = mirai::status(.compute = client$name)$daemons,
         shallow = FALSE
       )
       status <- unname(daemons[, "online", drop = TRUE])
@@ -158,7 +158,7 @@ crew_test("crew_launcher_local() can run a task and end a worker", {
   crew::crew_retry(
     ~{
       daemons <- rlang::duplicate(
-        x = mirai::daemons(.compute = client$name)$daemons,
+        x = mirai::status(.compute = client$name)$daemons,
         shallow = FALSE
       )
       status <- unname(daemons[, "online", drop = TRUE])
@@ -179,7 +179,7 @@ crew_test("crew_launcher_local() can run a task and end a worker", {
   crew::crew_retry(
     ~{
       daemons <- rlang::duplicate(
-        x = mirai::daemons(.compute = client$name)$daemons,
+        x = mirai::status(.compute = client$name)$daemons,
         shallow = FALSE
       )
       status <- unname(daemons[, "online", drop = TRUE])